Added support for listing the latest state of filesystems across XR, Admin and Host on Active and Standby. Also added support for collating and creating a tar ball out of syslogs across the system

+++ cleanup.sh
@@ -1,8 +1,16 @@
 #!/bin/bash
 
+set -x
+
+SCRIPT_PATH=$(dirname `which $0`)
+
+cd $SCRIPT_PATH
+
 
 # Clean up all the build and dist directories
 # top level + at the xr, admin and host level
 
-rm -r build/
-rm -r dist/
+if [[ -d build ||  -d dst ]]; then
+    rm -r build/
+    rm -r dist/
+fi
